requirements.txt
asn1crypto==1.5.1
blinker==1.7.0
cffi==1.16.0
click==8.1.7
coincurve==18.0.0
cryptography==41.0.7
cytoolz==0.12.2
eciespy==0.4.1
eth-hash==0.5.2
eth-keys==0.4.0
eth-typing==3.5.2
eth-utils==2.3.1
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pyasn1==0.5.1
pycparser==2.21
pycryptodome==3.19.0
toolz==0.12.0
typing_extensions==4.9.0
Werkzeug==3.0.1
